// src/app/models/FilterClass.model.ts
export class FilterClass {
    UserID: string;
    CurrentPage: number;
    Records: number;
    Status: string;
    StartDate: string;
    EndDate: string;
    InvoiceNumber: string;
    Organization: string;
    Division: string;
    Plant: string;
    PlantList: string[];
    CustomerName: string;
    LRNumber: string;
}